[redis-tls] disable tls verification for metrics
The redis exporter is unable to scrape metrics as it doesn't have a way of
verifying the server TLS cert.

Currently the chart sets the REDIS_EXPORTER_TLS_CA_CERT_FILE envvar but my
suspicion is that redis_exporter doesn't support TLS without client certs so it
tries to present a (presumably empty) client cert that the server then doesn't
like.

Switching off TLS verification seems to fix this.

The readiness and liveness checks still check TLS so we will still become aware
if the TLS cert is bad.  Also the exporter is scraping over localhost so I don't
think we're that concerned about TLS verification here.