[redis-tls] Manage Certificate
I keep using this component and forgetting to add and wire up a certificate to
it.  It's currently fiddly to do so because you have to patch the name of the
TLS secret to match the one used by the Certificate.  You can't use a standard
TLS secret name because each redis instance needs to use a different secret.

Unfortunately because this Secret isn't managed by Kustomize, by default it
won't get modified by namePrefix.  This commit therefore adds a namePrefix
configuration so that it will get updated when overlays use a namePrefix
transform.

This also updates the README to make clearer what the caller's responsibilities
are.