[redis-tls] reload TLS certs when they roll (#2)
Redis doesn't watch its cert files to see if they update.

This adds a simple CronJob that pokes redis every three hours to trigger it to
reload its TLS config if it has changed.  This allows us to change certificate
without downtime.  This is important in environments with 24h-lifetime certificates.